fix(components): remove back button rotation on `post-header` (#5466)
## 📄 Description

When the header was first developed, we didn't have as many icons
available and the `arrowleft` icon was missing. Therefore we had used
the `arrowright` icon with a 180 deg rotation on whatever the user was
putting in the back button slot.

Now that we have the correct icon, this PR removes the 180deg rotation
so that users can use the correct icon which is `arrowleft`.

---------

Co-authored-by: Alizé Debray <33580481+alizedebray@users.noreply.github.com>