chore(docs): include `raw-components` and `health` stories in the local build (#5377)
## 📄 Description
Some of the stories should only be shown when the docs are run locally.
Up to now, they were just hidden from the navigation sidebar using css.
That meant they could still be found and accessed by searching for them
in the searchbar.
To hide them not only from the sidebar but also the search a filter
function was added which checks for a custom `devOnly` tag which can be
added to any story. When the `data-env` attribute on the html tag is set
to production the filter is active and stories including the `devOnly`
tags are hidden. The attribute can be set to development in the devtools
to deactivate the filtering. All components will then be shown in the
sidebar and also searchable.

## 🚀 Demo

If applicable, please add a screenshot or video to illustrate the
changes.

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes

---------

Co-authored-by: Oliver Schürch <oliver.schuerch@post.ch>