fix(setup): versions.json updating within release workflow (#5666)
## 📄 Description
There was an issue in how the release workflow was updating the
versions.json. There are essentially two cases:

**Minor release** (e.g. 7.2 -> 7.3):
This case worked fine. The entry for version 7.x in the versions.json
was updated as well as the dependencies.

**Major releases** (e.g. 7.x -> 8.0):
Here is where the problem was. The code not only added a new entry for
the 8.0 version but also wrongfully updated the dependencies of the 7.x
entry to be the same as the 8.0 dependencies. That way every time a new
major release was done the dependencies of the previous major version
are messed up.

This pr fixes this. For major releases the dependencies of the previous
versions are left alone and the new entry is just added at the start. I
tried to test this by mocking the workflow to see how the versions.json
would look like and it seems correct. The fix will only adress future
releases so the current versions.json is still messed up. Thats why i
corrected it manually using the information form the packages.json
files.

## 🚀 Demo

If applicable, please add a screenshot or video to illustrate the
changes.

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes

---------

Co-authored-by: Alizé Debray <33580481+alizedebray@users.noreply.github.com>