chore(docs): generate dependency table for styles package dynamically from `versions.json` (#5669)
## 📄 Description
This only implements part of the ticket. The information available in
the `versions.json` does not allow us to dynamically generate the
current release line table so only the dependencies table is filled
dynamically. For this a small react component is introduced which reads
uses the contents of the `versions.json` file, formats the versions and
generates the table.

## 🚀 Demo
Before:

![image](https://github.com/user-attachments/assets/97749d15-a2eb-4af3-8e3c-48e8ee6130dd)

After:

![image](https://github.com/user-attachments/assets/c8ecc2c4-9a48-4f84-b4e4-147f99e4d57b)

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes

---------

Co-authored-by: Alizé Debray <33580481+alizedebray@users.noreply.github.com>