chore(components): update language switch accessible names and roles (#5350)
## 📄 Description
This PR initial goal was to fix non conformity with the
https://www.w3.org/WAI/WCAG22/Understanding/label-in-name accessibility
guideline. Instead of modifying the `aria-label` content this PR
replaces the `aria-labels` set on the contained button elements with
descriptive visually hidden span texts. List variant 'role' attrs are
modified to correctly recognize list items. An info box is added in the
language switch page of the documentation to inform about proper usage
of the `name` language option prop.

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- ✔️ New and existing unit tests pass locally with my changes

---------

Co-authored-by: Lea <lea.gardavaud@post.ch>