feat(components): add default styles to the post-menu (#5632)
## 📄 Description

This PR is to set default styles to the `post-menu` component matching
the designed menu.
**At the moment the `post-menu` remains a raw component and is not yet
public, this is why no changeset was added.**

This PR will be followed by other PRs in order to update the component's
logic and then make it publicly available by publishing a documentation
page.

## 🚀 Demo

If applicable, please add a screenshot or video to illustrate the
changes.

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes