chore(components): improved URL handling for `post-icon` component (#5109)
## 📄 Description

Improved URL handling in `post-icon` component:

I changed the order of which "sources" are used to determine the origin
and the slug. Now the order is the following: base property > base tag >
data-post-icon-base meta attribute. This means that if we have both a
base tag and a meta attribute present the url built using the
information in the base tag first and only if either the origin/slug
part are missing it takes them from the meta attribute.
I also changed how a root path ("/") is interpreted. Before it would not
count as a slug and would therefore be ignored.

## 🚀 Demo

If applicable, please add a screenshot or video to illustrate the
changes.

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes