chore(components): create eslint rule for stencil props initialization (#5505)
## 📄 Description

This PR introduces a custom ESLint rule that checks for any Stencil
component @Prop property that lacks an initial value, is not marked as
optional (?), and is not explicitly marked as definitely assigned (!).
It also adds `design-system-eslint` as a dependency to the components
package.

Must be reviewed after
[#4651](https://github.com/swisspost/design-system/issues/4651) is
merged.

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes