chore(documentation): required and optional fields on form elements (#5622)
## 📄 Description

Added the documentation for the required and optional fields. As decided
with the team (see ticket's comments), the arguments were added to each
form element but the general documentation was put in "Form creation"
guidelines.

The required/optional were added separately from the valid/invalid
arguments as a field can be required/optional as well as valid/invalid
so it did not make much sense to me to have them together.

Also fixed a small issue in the documentation were the slider
valid/invalid feedback didn't have an `id` and was therefore not
connected to the `aria-describedby` of the input.

---------

Co-authored-by: léa <lea.gardavaud@gmail.com>