chore(styles, docs): remove bootstrap dependency (#5850)
## 📄 Description

Removed the bootstrap package from the dependencies of the `styles`,
`documentation` and `internet-header` packages.
Removed all mentions of bootstrap within the docs as well as all imports
of the bootstrap css in the `styles` package.

Some broken components were fixed in this PR (`table`, `product-card`
and `button-group`) by internalizing the necessary code from bootstrap.

I will create tickets for the two remaining ones: `Slider` and `Card
footer/header`.

The PR #5857 handles the removal of all of the variable overrides in the
`styles` package that are no longer needed nor used, as well as the
deletion of the slider (form range) component.

---------

Co-authored-by: Alizé Debray <33580481+alizedebray@users.noreply.github.com>