chore: update storybook to 9 (#5655)
## 📄 Description

Updates storybook to the new major version of storybook, v9. All
suggested auto updates have been applied, thus the many changes in the
files.

Known bugs:
- There is a react error connected to the `sourceState="shown"`
attribute. If it's not there, code samples are not open by default and
the console error is gone. If the attribute is set, the error appears in
the console but the page generally seems to work. Further investigation
is necessary, but the issue does not seem blocking.


---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes