chore(docs): update version file (#6008)
## 📄 Description

- Removed the `versions.json` file.
- Since we store the `versions.json` file only on the latest release
branch (release/v9) and on main, we must make sure not to use it
locally, when we release a next major version. Therefore, the
dependencies table component on the styles package page had to be
updated.
- The tokens build script did not handle multi-value tokens (except from
type `typography` tokens) well, which led to CI errors sometimes. I've
copied over the files we have on main and just updated the constants.js
file, to match the current setup.

---

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes

---------

Co-authored-by: Oliver Schürch <oli@skaos.ch>