feat(docs, styles): improve position utilities. (#6002)
## 📄 Description

The _position_ utility classes allow defining the inset of an element.
The classes available are
- top-[position]
- bottom-[position]
- start-[position]
- end-[position]

where [position] can be 0, 50, or 100.

The goal was to improve these _position_ utilities and their
documentation by:

- `inline-*` and `block-*` properties: Replaced `top`, `bottom`, `right`
and `left` with
`inset-block-start`, `inset-block-end`, `inset-inline-start` and
`inset-inline-end`.
- Added a control for the `bottom-*` and `end-* classes` in the
_position_ utility documentation.

Consequently classes for `bottom` and `end` were added to the SCSS
variables on styles.

---------

Co-authored-by: Alona Zherdetska <138328641+alionazherdetska@users.noreply.github.com>