feat(eslint): add automigration rules to update spacing utilities (#6219)
## 📄 Description

Add auto migration rules for the spacing utility classes.
Migration is done in two phases to avoid errors due to classes that are
named the same before and after the migration (e.g. deprecated class
`mt-4` which was equal to `margin-top: 24px` and new class `mt-4` which
is equal to `margin-top: 4px`)

---------

Co-authored-by: Alizé Debray <alize.debray@post.ch>