fix(styles): remove user-menu need for a slot attribute (#6460)
## 📄 Description

This PR update the user menu styles so that it does not need to be
wrapped in a element with a `slot="global-login"` attribute when it's
not in the global header (i.e. when it is in the custom controls).

## 🚀 Demo

If applicable, please add a screenshot or video to illustrate the
changes.

---

## 🔮 Design review

- [ ] Design review done
- [x] No design review needed

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes