fix(components): add megadropdown overview link to the focusable elements (#6557)
## 📄 Description

When navigating with keyboard navigation on the megadropdown, the newly
added **overview link** was not getting focused as it had not been added
to the loop of focusable elements.
This PR adds it at the start of the list so that if present, it becomes
the first focusable element.

---

## 🔮 Design review

- [ ] Design review done
- [X] No design review needed

## 📝 Checklist

- ✅ My code follows the style guidelines of this project
- 🛠️ I have performed a self-review of my own code
- 📄 I have made corresponding changes to the documentation
- ⚠️ My changes generate no new warnings or errors
- 🧪 I have added tests that prove my fix is effective or that my feature
works
- ✔️ New and existing unit tests pass locally with my changes