chore(docs): add label to components in the next docs side-nav (#6044)
## 📄 Description

Added label to indicate **status** of components in **foundations**,
**raw components**, **components**, **utilities**, **templates** on
documentation.

**Status label** indicates the _state of the component_:
- **In Progress ⏳:** This component has open issues or tasks that need
to be fixed before it can be used.
- **Experimental 🆕:** This component has been newly introduced and you
might experience some bumps while using it.
- **Stable ✅:** This component is tested, has some mileage, other
projects are already successfully implementing it, it's safe to use
this.
- **Deprecated ⛔:** This component will be removed in an upcoming major
version.

Changed default svg-icons to be hidden in `production` and in
`development` view in styles.

---------

Co-authored-by: Alona Zherdetska <138328641+alionazherdetska@users.noreply.github.com>
Co-authored-by: leagrdv <lea.gardavaud@post.ch>